Fix sync bugs, multiplayer movement now works

+++ src/lib-tempo/include/tempo/network/base.hpp
@@ -76,12 +76,12 @@ namespace tempo
 
 	static sf::IpAddress addr_l = sf::IpAddress::getLocalAddress();
 	extern sf::IpAddress addr_r;  // Remote Address
-	extern unsigned int port_ci;  // Client Incoming Port
-	extern unsigned int port_co;  // Client Outgoing Port
-	extern unsigned int port_sh;  // Server Hand Shake Port (Primary Port)
-	extern unsigned int port_si;  // Server Incoming Port
-	extern unsigned int port_so;  // Server OutGoing Port
-	extern unsigned int port_st;  // Server Time Sync Port
+	extern unsigned short port_ci;  // Client Incoming Port
+	extern unsigned short port_co;  // Client Outgoing Port
+	extern unsigned short port_sh;  // Server Hand Shake Port (Primary Port)
+	extern unsigned short port_si;  // Server Incoming Port
+	extern unsigned short port_so;  // Server OutGoing Port
+	extern unsigned short port_st;  // Server Time Sync Port
 
 	////////////////////////////////////////////////////////////////////////
 	/// Function Declarations
